Move DurationParser and tests into Abstractions
