+++ NuKeeper.Abstractions.Tests/Formats/DurationParserTests.cs
@@ -1,7 +1,8 @@
 using System;
+using NuKeeper.Abstractions.Formats;
 using NUnit.Framework;
 
-namespace NuKeeper.Tests
+namespace NuKeeper.Abstractions.Tests.Formats
 {
     [TestFixture]
     public class DurationParserTests
